Added item delete command
I think I removed the item from all of the contexts that it needs to
be removed from...there's not really a good way to test this yet
because I have no way of sharing items at the moment. I just wanted to work
on this a little before going to bed.