Not a back-tracker (!)
Added command to trigger list generation...eventually it will be automatic,
but that seems unnecessary at the moment.

Can determine which items need to be bought, but nothing gets
added to a user's list yet. Maybe I'll do that after lunch. Or maybe I'll
go biking.

Having a backtracker would be nifty, but might be overkill