Paving the way for making the list page actually work
I added a command to display what's on a user's list, which will be very
helpful when I start testing the list page.

I fixed a whole ton (20 or so) of warnings with my trusty @SuppressWarnings
annotation. Calm down--most of them were because I didn't use a
TemplateEngine in some routes (like POSTSignOutRoute).
In some cases, the warnings were actually because I did something stupid,
so I actually fixed those.

I also implemented Serializable in the PQComparator class, which was preventing
anything from being read in at startup haha

I finalized (maybe) the freemarker template. I won't know if it's right until
I finish the GETListRoute.

I also found a bug with list generation:
The quantity a user is responsible to buy does not actually increase.
For example, the logs show:
    jonny_depp is responsible for buying 1 Gal Milk
    ...later...
    jonny_depp is repsonsible for buying 1 Gal Milk

...but jonny_depp's list only shows 1 Gal of Milk total

I don't know if that made any sense, but I'm too tired to read it over.
SEND IT!