Serilization Sucks
I'm having a lot of trouble making a hashCode() method that returns
the same thing before and after the fridge is serialized.

I already can't wait to rewrite this with a database.

It's currently in a semi-stable state
Presonal Items are persisted, but Shared Items don't always persist
if the server goes down.

There are some scary nulls that I don't like, but aren't really causing
that many problems. In the interest of finishing this quickly...what nulls?